requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
executing==2.2.0
flash==1.0.3
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
polars==1.31.0
prettyprint==0.1.5
protobuf==6.31.1
pydeck==0.9.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wsproto==1.2.0
